requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
APScheduler==3.10.4
python-dotenv==1.0.0
pydantic==2.5.0
slowapi==0.1.9
pytz==2024.1
requests==2.31.0

